Add files via upload
Suppose you’re selling 100 toffees. Every time a customer comes, he buys random number of toffees. Create a function which deducts the number of toffees bought by user. This number of toffees should be as an input from user. Keep selling the toffees until it is out of stock. Store the transaction record of number of toffees sold every time a customer came, in a list